r.resamp.filter, r.resamp.interp: fix integer overflow

+++ raster/r.resamp.interp/main.c
@@ -246,10 +246,10 @@ int main(int argc, char *argv[])
 
                         if (Rast_is_d_null_value(&c)) {
                             Rast_set_d_null_value(
-                                &outbuf[(row - start) * dst_w.cols + col], 1);
+                                &outbuf[(size_t)(row - start) * dst_w.cols + col], 1);
                         }
                         else {
-                            outbuf[(row - start) * dst_w.cols + col] = c;
+                            outbuf[(size_t)(row - start) * dst_w.cols + col] = c;
                         }
                     }
 
@@ -288,10 +288,10 @@ int main(int argc, char *argv[])
                             Rast_is_d_null_value(&c10) ||
                             Rast_is_d_null_value(&c11)) {
                             Rast_set_d_null_value(
-                                &outbuf[(row - start) * dst_w.cols + col], 1);
+                                &outbuf[(size_t)(row - start) * dst_w.cols + col], 1);
                         }
                         else {
-                            outbuf[(row - start) * dst_w.cols + col] =
+                            outbuf[(size_t)(row - start) * dst_w.cols + col] =
                                 Rast_interp_bilinear(u, v, c00, c01, c10, c11);
                         }
                     }
@@ -361,10 +361,10 @@ int main(int argc, char *argv[])
                             Rast_is_d_null_value(&c32) ||
                             Rast_is_d_null_value(&c33)) {
                             Rast_set_d_null_value(
-                                &outbuf[(row - start) * dst_w.cols + col], 1);
+                                &outbuf[(size_t)(row - start) * dst_w.cols + col], 1);
                         }
                         else {
-                            outbuf[(row - start) * dst_w.cols + col] =
+                            outbuf[(size_t)(row - start) * dst_w.cols + col] =
                                 Rast_interp_bicubic(u, v, c00, c01, c02, c03,
                                                     c10, c11, c12, c13, c20,
                                                     c21, c22, c23, c30, c31,
@@ -419,7 +419,7 @@ int main(int argc, char *argv[])
                         }
 
                         if (do_lanczos) {
-                            outbuf[(row - start) * dst_w.cols + col] =
+                            outbuf[(size_t)(row - start) * dst_w.cols + col] =
                                 Rast_interp_lanczos(u, v, c);
                         }
                     }
@@ -433,7 +433,7 @@ int main(int argc, char *argv[])
 
         /* write to output map */
         for (row = start; row < end; row++) {
-            Rast_put_d_row(outfile, &outbuf[(row - start) * dst_w.cols]);
+            Rast_put_d_row(outfile, &outbuf[(size_t)(row - start) * dst_w.cols]);
         }
         written = end;
     }
